Migrate to Makefile build system
The standard Arduino project layout doesn't play well with VCS due to
its use of ZIP files for library distribution. Makefiles are much more
standard and flexible in that domain, although they are not portable.
However, multi-platform support is not really necessary considering this
project is mainly for private use.